+++ src/main/java/kitchenpos/application/OrderCreationValidator.java
@@ -1,0 +1,7 @@
+package kitchenpos.application;
+
+import kitchenpos.domain.Order;
+
+public interface OrderCreationValidator {
+    void validate(final Order order);
+}
