--- src/main/java/kitchenpos/domain/OrderValidator.java
@@ -1,5 +1,0 @@
-package kitchenpos.domain;
-
-public interface OrderValidator {
-    void validate(final Order order);
-}
